Convert mutation-observer.browser.js
